Added DriveForATime that drives the robot for a fixed length of time at a fixed speed. Added DriveOffPlatform that will enable the vacuums to hold a game piece, then drive off the platform at a good speed to land well. Updated drive to use doubles instead of floats (to reduce the number of casts when dealing with drive)

+++ src/main/java/frc/robot/commands/autoassistAlignment.java
@@ -94,7 +94,7 @@ public class autoassistAlignment extends Command {
     // and direction you are facing
 
     double throttleJoystick = Robot.m_oi.driverController.getRawAxis(RobotMap.driverControllerAxisFrontAndBack);
-    Robot.m_drive.driveDirection((float) throttleJoystick, (float) m_output);
+    Robot.m_drive.driveDirection(throttleJoystick, m_output);
 
     SmartDashboard.putNumber("autoAssistdriveOutput", m_output);
